remove extra files, condense everything to WorkcamKeybindActions
